# agent/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Autonomous Hydroponic Plant API",
    "description": "REST API to control and monitor the ZombiePlant V1.0 hydroponic system.",
    "version": "1.1"
  },
  "paths": {
    "/tools/feed": {
      "post": {
        "tags": [
          "Tools"
        ],
        "summary": "Smart Feed Cycle",
        "description": "Performs a complete drain-refill-mix cycle with nutrient dosing.\n1. Empty Tank.\n2. Dose Nutrients (Micro/Gro/Bloom).\n3. Fill to Max with fresh water.\n4. Mix (Air Stones/Light ON) for 3 minutes.\n5. Verify TDS.",
        "operationId": "smart_feed_cycle_tools_feed_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tools/dose": {
      "post": {
        "tags": [
          "Tools"
        ],
        "summary": "Smart Dose",
        "description": "Precise nutrient addition without draining the tank.\n1. Dispenses specific nutrient amount.\n2. Mixes (Air Stones) for specified duration.\n3. Returns new TDS reading.",
        "operationId": "smart_dose_tools_dose_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DoseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DoseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tools/flush": {
      "post": {
        "tags": [
          "Tools"
        ],
        "summary": "True System Flush",
        "description": "Performs a full system flush (Rinse Cycle).\n1. Empty Tank.\n2. Fill to Max (Fresh Water).\n3. Soak/Mix (Air Stones ON) for X seconds.\n4. Empty Tank (Drain Rinse).\n5. Fill to Max (Fresh Water).",
        "operationId": "true_system_flush_tools_flush_post",
        "parameters": [
          {
            "name": "soak_duration",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Duration in seconds to run the air stones/mix cycle.",
              "default": 180,
              "title": "Soak Duration"
            },
            "description": "Duration in seconds to run the air stones/mix cycle."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FlushResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/jobs/": {
      "post": {
        "tags": [
          "Jobs"
        ],
        "summary": "Submit Job",
        "description": "Submit a new background job.\nReturns the initial job status (queued).",
        "operationId": "submit_job_jobs__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatus"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/jobs/{job_id}": {
      "get": {
        "tags": [
          "Jobs"
        ],
        "summary": "Get Job Status",
        "description": "Get the status of a specific job.",
        "operationId": "get_job_status_jobs__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The ID of the job to retrieve",
              "title": "Job Id"
            },
            "description": "The ID of the job to retrieve"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatus"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Jobs"
        ],
        "summary": "Cancel Job",
        "description": "Cancel a running or queued job.",
        "operationId": "cancel_job_jobs__job_id__delete",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DiagnosticResponse": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "healthy",
              "warning",
              "error"
            ],
            "title": "Status"
          },
          "sensors": {
            "additionalProperties": {
              "$ref": "#/components/schemas/SensorCheckResult"
            },
            "type": "object",
            "title": "Sensors"
          },
          "pumps": {
            "$ref": "#/components/schemas/PumpCheckResult"
          }
        },
        "type": "object",
        "required": [
          "status",
          "sensors",
          "pumps"
        ],
        "title": "DiagnosticResponse"
      },
      "DoseRequest": {
        "properties": {
          "nutrient": {
            "$ref": "#/components/schemas/PumpID",
            "description": "The nutrient pump to activate (flora_micro, flora_gro, flora_bloom)."
          },
          "amount_ml": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Amount Ml",
            "description": "Amount of nutrient to dispense in mL."
          },
          "mix_seconds": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Mix Seconds",
            "description": "Duration to run air stones for mixing after dosing.",
            "default": 30
          }
        },
        "type": "object",
        "required": [
          "nutrient",
          "amount_ml"
        ],
        "title": "DoseRequest"
      },
      "DoseResponse": {
        "properties": {
          "status": {
            "type": "string",
            "const": "success",
            "title": "Status",
            "default": "success"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "nutrient": {
            "$ref": "#/components/schemas/PumpID"
          },
          "amount_dispensed_ml": {
            "type": "number",
            "title": "Amount Dispensed Ml"
          },
          "final_tds": {
            "type": "number",
            "title": "Final Tds"
          }
        },
        "type": "object",
        "required": [
          "message",
          "nutrient",
          "amount_dispensed_ml",
          "final_tds"
        ],
        "title": "DoseResponse"
      },
      "EmptyResponse": {
        "properties": {
          "status": {
            "type": "string",
            "const": "success",
            "title": "Status",
            "default": "success"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "duration": {
            "type": "number",
            "title": "Duration"
          }
        },
        "type": "object",
        "required": [
          "message",
          "duration"
        ],
        "title": "EmptyResponse"
      },
      "FeedRequest": {
        "properties": {
          "recipe": {
            "$ref": "#/components/schemas/NutrientRecipe",
            "description": "Nutrient recipe to apply."
          },
          "amounts_ml": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "number"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amounts Ml",
            "description": "Custom nutrient amounts in mL (required if recipe is custom). Keys: micro, gro, bloom."
          }
        },
        "type": "object",
        "required": [
          "recipe"
        ],
        "title": "FeedRequest"
      },
      "FeedResponse": {
        "properties": {
          "status": {
            "type": "string",
            "const": "success",
            "title": "Status",
            "default": "success"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "recipe": {
            "$ref": "#/components/schemas/NutrientRecipe"
          },
          "amounts_dispensed": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Amounts Dispensed"
          },
          "final_tds": {
            "type": "number",
            "title": "Final Tds"
          }
        },
        "type": "object",
        "required": [
          "message",
          "recipe",
          "amounts_dispensed",
          "final_tds"
        ],
        "title": "FeedResponse"
      },
      "FillResponse": {
        "properties": {
          "status": {
            "type": "string",
            "const": "success",
            "title": "Status",
            "default": "success"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "fill_duration": {
            "type": "number",
            "title": "Fill Duration"
          },
          "adjust_duration": {
            "type": "number",
            "title": "Adjust Duration"
          }
        },
        "type": "object",
        "required": [
          "message",
          "fill_duration",
          "adjust_duration"
        ],
        "title": "FillResponse"
      },
      "FlushResponse": {
        "properties": {
          "status": {
            "type": "string",
            "const": "success",
            "title": "Status",
            "default": "success"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "final_fill_details": {
            "$ref": "#/components/schemas/FillResponse"
          }
        },
        "type": "object",
        "required": [
          "message",
          "final_fill_details"
        ],
        "title": "FlushResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JobRequest": {
        "properties": {
          "type": {
            "$ref": "#/components/schemas/JobType"
          },
          "params": {
            "anyOf": [
              {
                "additionalProperties": {
                  "anyOf": [
                    {
                      "type": "string"
                    },
                    {
                      "type": "number"
                    },
                    {
                      "type": "integer"
                    },
                    {
                      "additionalProperties": true,
                      "type": "object"
                    }
                  ]
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Params",
            "description": "Parameters for the job (e.g., recipe for feed)."
          }
        },
        "type": "object",
        "required": [
          "type"
        ],
        "title": "JobRequest"
      },
      "JobState": {
        "type": "string",
        "enum": [
          "queued",
          "running",
          "completed",
          "failed"
        ],
        "title": "JobState"
      },
      "JobStatus": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "type": {
            "$ref": "#/components/schemas/JobType"
          },
          "status": {
            "$ref": "#/components/schemas/JobState"
          },
          "created_at": {
            "type": "number",
            "title": "Created At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "string"
              },
              {
                "$ref": "#/components/schemas/FillResponse"
              },
              {
                "$ref": "#/components/schemas/EmptyResponse"
              },
              {
                "$ref": "#/components/schemas/FlushResponse"
              },
              {
                "$ref": "#/components/schemas/FeedResponse"
              },
              {
                "$ref": "#/components/schemas/DiagnosticResponse"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "type",
          "status",
          "created_at"
        ],
        "title": "JobStatus"
      },
      "JobType": {
        "type": "string",
        "enum": [
          "fill_to_max",
          "empty_tank",
          "system_flush",
          "feed",
          "diagnose"
        ],
        "title": "JobType"
      },
      "NutrientRecipe": {
        "type": "string",
        "enum": [
          "vegetative",
          "flowering",
          "custom"
        ],
        "title": "NutrientRecipe"
      },
...
      "PumpCheckResult": {
        "properties": {
          "passed": {
            "type": "boolean",
            "title": "Passed"
          },
          "pump_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pump Id"
          },
          "noise_level": {
            "type": "number",
            "title": "Noise Level"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "passed",
          "noise_level",
          "message"
        ],
        "title": "PumpCheckResult"
      },
...
      "PumpID": {
        "type": "string",
        "enum": [
          "water_out",
          "water_in",
          "flora_micro",
          "flora_gro",
          "flora_bloom"
        ],
        "title": "PumpID"
      },
...
      "SensorCheckResult": {
        "properties": {
          "passed": {
            "type": "boolean",
            "title": "Passed"
          },
          "value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "string"
              }
            ],
            "title": "Value"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "passed",
          "value",
          "message"
        ],
        "title": "SensorCheckResult"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}